+++ ChangeLog.md
@@ -2,9 +2,13 @@
   * Update documentation for material and rarity information
   * Update coriolis export for Beluga Liner
   * Help large star system databases by adding an index
-  * Add 'shared' flag for bounty events
-  * Add 'source' for ship refuelled
+  * Add "shared" flag for bounty events
+  * Add "source" for ship refuelled
   * Ship refuelled event now triggered when finishing scooping fuel
+  * Handle renamed fields for system information in FSD jumps in 2.2.0.2
+  * Add "Modification crafted" event
+  * Add "Modification applied" event
+  * Add "Engineer progressed" event
 
 #2.0.9
   * Do not update ship configuration when data is not available from the companion API